requirements.txt
flask
flask_session
flask_socketio
cachelib
boto3
openai
google-generativeai